[clean] controller service refactoring
