+++ server/api/user/_userId@string/relationships/controller.ts
@@ -1,10 +1,6 @@
-import {
-  followers,
-  createRelationship,
-  unfollowUser,
-  indexRelationship
-} from '$/service/user'
 import { defineController } from './$relay'
+
+import { indexRelationship } from '$/service/relationship'
 
 export default defineController(() => ({
   get: async ({ currentUserUid }) => {
@@ -12,31 +8,6 @@ export default defineController(() => ({
       const users = await indexRelationship(currentUserUid)
 
       return { status: 200, body: { users } }
-    } catch (error) {
-      return { status: 500, body: { error } }
-    }
-  },
-  post: async ({ params, currentUserUid }) => {
-    try {
-      const userId = params.userId
-
-      await createRelationship(userId, currentUserUid)
-
-      return { status: 201, body: { message: 'フォロー申請を送りました。' } }
-    } catch (error) {
-      return { status: 500, body: { error } }
-    }
-  },
-  delete: async ({ params, currentUserUid }) => {
-    try {
-      const userId = params.userId
-
-      await unfollowUser(userId, currentUserUid)
-
-      return {
-        status: 201,
-        body: { message: 'フォロー申請を取り下げました。' }
-      }
     } catch (error) {
       return { status: 500, body: { error } }
     }
