+++ server/service/user.ts
@@ -17,7 +17,7 @@ export const indexUser = async (take: number, skip: number) => {
 /**
  * show
  */
-export const showUser = async (userId: string, currentUserUid: string) => {
+export const showUser = async (userId: string) => {
   const user = await prisma.user.findUnique({
     where: { id: userId }
   })
@@ -40,82 +40,4 @@ export const createUser = async (body: UserCreateBody) => {
   })
 
   return user
-}
-
-/**
- * followers
- */
-export const followers = async (userId: string) => {
-  const user = await prisma.user.findUnique({
-    where: { id: userId },
-    include: { followed: {} }
-  })
-  if (!user)
-    throw Object.assign(new Error('ユーザーが存在しません。'), { status: 404 })
-
-  return user.followed
-}
-
-export const indexRelationship = async (currentUserUid: string) => {
-  const user = await prisma.user.findUnique({
-    where: { id: currentUserUid },
-    include: { following: true }
-  })
-
-  if (!user)
-    throw Object.assign(new Error('ユーザーが存在しません。'), { status: 404 })
-
-  return user.following
-}
-
-export const isRelationship = async (
-  userId: string,
-  currentUserUid: string
-) => {
-  const user = await prisma.user.findUnique({
-    where: { id: userId },
-    include: {
-      followed: {
-        where: { id: currentUserUid }
-      }
-    }
-  })
-
-  if (!user)
-    throw Object.assign(new Error('ユーザーが存在しません。'), { status: 404 })
-
-  const bool = user.followed.length === 0 ? false : true
-
-  return bool
-}
-
-/**
- * follow
- */
-export const createRelationship = async (
-  userId: string,
-  otherUserId: string
-) => {
-  await prisma.user.update({
-    where: { id: userId },
-    data: {
-      followed: {
-        connect: { id: otherUserId }
-      }
-    }
-  })
-}
-
-/**
- * unfollow
- */
-export const unfollowUser = async (userId: string, currentUserUid: string) => {
-  await prisma.user.update({
-    where: { id: currentUserUid },
-    data: {
-      following: {
-        disconnect: { id: userId }
-      }
-    }
-  })
 }
